requirements.txt
django-pagedown
asgiref==3.2.7
boto3==1.12.47
botocore==1.15.47
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1